codex - add reusable docs drawer
